Prefer builtins.sum over sum
